Added rudimentary ability activation
Hopefully will have something more robust in the future (And something that doesn't look like crap) but for now this is a good start. Things to do for this change in the near future: Make it not look crap (Topest priority) Give users the ability to add resistances that add resistance roll chat prompts. Those prompts should roll the appropriate skill/defense and display the result as well as degrees of success/failure, but all of that is for another time, this is just the basics of getting abilities to interact with the chat.